CARDS-1245: We should allow project-specific frontend code
