+++ aggregated-frontend/pom.xml
