+++ aggregated-frontend/src/main/frontend/webpack.config-template.js
