+++ aggregated-frontend/src/main/resources/.gitignore
